requirements.txt
tweepy==4.14.0
chromadb==0.4.15
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
umap-learn==0.5.3
hdbscan==0.8.33
python-dotenv==1.0.0